refator the code
